backend/requirements.txt
# Core Django
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database (PostgreSQL - opcional para producción)
# Para desarrollo usaremos SQLite (incluido en Python)
# Para producción con PostgreSQL, instalar: pip install psycopg[binary]
# psycopg[binary]>=3.1.0

# WebSockets for real-time chat
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# CORS headers
django-cors-headers==4.3.1

# Image handling
Pillow>=11.0.0

# Environment variables
python-decouple==3.8

# API Documentation (optional but recommended)
drf-spectacular==0.27.1

# Development tools
pytest==8.0.0
pytest-django==4.8.0

